Show correct embed URL
